Added pitchSelect() and hitSelect()
pitchSelect is what is called when the user is pitching and hitSelect is
what is called when the user is hitting.
This is where animations and sound effects can be played.
I'll work on getting functions that can be linked to the buttons in the
html page that can link to these functions.